Revert to analog read for EOL_R
EOL_R should be a digital signal (with a pull-up resistor enabled on the MCU
and driven by an open-drain diode on the Brother side, but it seems the high
level is close to the mid level [for some machines?] for reasons still unknown).

The conversion from analog to digital reading for this input was introduced
with the implementation of automatic shield correction detection relative
to the L input of the KH910's right-hand sensor.